feat: refactor podcast service
