+++ src/services/Podcast.service.ts
@@ -3,17 +3,17 @@ import { API } from "utils/constants/api.constants"
 import { IPodcastDetails } from "utils/interfaces/podcast-detail.interface"
 import { IPodcasts } from "utils/interfaces/podcasts.interface"
 
-const podcastService = () => {
-  const getPodcasts = async (): Promise<IPodcasts> => {
+const podcastService = {
+  getPodcasts: async (): Promise<IPodcasts> => {
     return http
       .GET({ path: API.PODCASTS })
       .then((response) => response as IPodcasts)
       .catch((error) => {
         throw error
       })
-  }
+  },
 
-  const getPodcastDetails = async (podcastId: string): Promise<IPodcastDetails> => {
+  getPodcastDetails: async (podcastId: string): Promise<IPodcastDetails> => {
     return http
       .GET({ path: API.PODCAST_DETAIL(podcastId) })
       .then((response) => response as IPodcastDetails)
@@ -21,8 +21,6 @@ const podcastService = () => {
         throw error
       })
   }
-
-  return { getPodcasts, getPodcastDetails }
 }
 
-export default podcastService()
+export default podcastService
